src/auth/requirements.txt
bcrypt==3.2.0
blinker==1.8.2
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
cryptography==3.4.8
distro==1.7.0
fasteners==0.14.1
Flask==3.0.3
Flask-MySQLdb==2.0.0
future==0.18.2
httplib2==0.20.2
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==8.10.0
mysqlclient==2.2.4
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.8.1
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==3.0.4
xdg==5
zipp==1.0.0
